requirements.txt
aiogram==2.17.1
python-dateutil==2.8.2
oauth2client==4.1.3
google-api-python-client==2.33.0
humanfriendly==10.0

